Created about directory for AboutScreen files. Moved LibraryDefaults, interfaces, and data classes into their own file in order to clean up AboutScreen and because they will be used in separate screen. Created copy of LibraryDefaults, interfaces, and data classes that will be used by AboutScreen.
