+++ app/src/main/java/com/heyzeusv/androidutilitieslibrary/AboutScreen.kt
@@ -2,8 +2,8 @@ package com.heyzeusv.androidutilitieslibrary
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.heyzeusv.androidutilities.compose.ui.AboutScreen
-import com.heyzeusv.androidutilities.compose.ui.LibraryItem
+import com.heyzeusv.androidutilities.compose.ui.about.AboutScreen
+import com.heyzeusv.androidutilities.compose.ui.about.LibraryItem
 import com.heyzeusv.androidutilitieslibrary.ui.theme.AndroidUtilitiesLibraryTheme
 import com.heyzeusv.androidutilitieslibrary.util.fakeLibrary
